Added adjustment to renewal index
This better accounts for differences between the early and later parts of the season, such as shorter episode orders earlier on that might otherwise be seen as having a negative effect, which should only happen once other shows start getting a higher order.

This adjusts the average renewal index for each season to match the overall weighted average renewal index for all renewed/canceled shows on the network. By the end of the season, this factor will usually be very close to 1, and won't affect much, but it will be a big help towards the beginning.