Optimizing
Some parallel processing was slower than non parallel, and the other way around. Did some testing of timing and found these to be the optimal versions.

It's possible with more shows in the future that this will need to be re-examined, but for now this will perform slightly faster. Some non-parallel loops ran dozens of times faster than the parallel variants.