Changed Accuracy Calculation
Straight weighted accuracy is checkedc first,

then an "error" value is checked.

Error represents the total weighted difference between the calculated odds for the show, and 50%, calculated only on shows that were predicted incorrectly.

It also includes error values for every "Renewed for Final Season" show, based on the difference from 55%, as that is the ideal odds for a show in that position.

Ultimately, this should provide a slightly better error reduction while also giving an indication as to how much error must be reduced to correctly predict another show.