Added Functionality for Viewer Numbers
As Nielsen ratings become smaller and smaller, I felt it would be important to log total viewer numbers for all shows in the database, which the Neural model will eventually be able to use and determine what portion of viewer numbers vs ratings to use. Some shows may rely less on demographic ratings and some more. This will also allow for more precise ranking of shows with ratings very close to each other.