New Calculation for Prediction Confidence
The confidence of the prediction is inversely proportional to the overlap between the normal distribution of the projected rating and the normal distribution of the renewal threshold. If there's a lot of overlap, the prediction can not be very confident. If there's little overlap, the prediction can be very confident. This will help differentiate the confidence of predictions later in the season vs earlier. The previous method helped with this, but this will be more statistically accurate.